requirements.txt
djangorestframework==3.14.0
django==4.1.1
Pillow==9.2.0
djangorestframework_simplejwt==5.2.1
google-auth==2.13.0
psycopg2==2.9.3
requests==2.28.0
